backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx
redis
python-dotenv
